# backend/api/openapi_0_2.yaml
openapi: 3.1.0
info:
  title: HomePlanner API v0.2
  description: >
    Спецификация HTTP API для релиза HomePlanner v0.2.
    Все временные поля передаются в локальном формате проекта без таймзоны (`YYYY-MM-DD HH:MM:SS`).
  version: 0.2.0
servers:
  - url: /api/v0.2
    description: Базовый префикс API v0.2
...
paths:
...
  /tasks/stream:
    get:
      tags: [tasks]
      summary: WebSocket-стрим задач
      description: >
        WebSocket-канал для получения событий задач и напоминаний в реальном времени.
        Соединение устанавливается по протоколу `wss`, токен передаётся в query `token`.
      security:
        - bearerAuth: []
      responses:
        '101':
          description: Успешное переключение протокола на WebSocket
      x-websocket: true
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...